feat(GroupPurchaseRepositoryImpl.searchGroupPurchases): 공동구매 다건 조회 api 작업 완료 - 정렬 구현

+++ src/main/java/wj/flab/group_wise/dto/groupPurchase/response/GroupPurchaseSummaryResponse.java
@@ -12,14 +12,12 @@ public record GroupPurchaseSummaryResponse (
     GroupPurchase.Status status,
     int discountRate,
     int cheapestPrice,
-    long productStockCount,
+    int productStockCount,
     int minimumParticipants,
-    int currentParticipantCount,
-    int participationRate,
+    long currentParticipantCount,
+    double participationRate,
     long wishlistCount,
     LocalDateTime startDate,
     LocalDateTime endDate,
     Duration remainingTime
-){
-
-}
+){}
